+++ svc/pkg/captcha/ops/hcaptcha-verify/src/lib.rs
@@ -17,7 +17,11 @@ async fn handle(
 ) -> GlobalResult<captcha::hcaptcha_verify::Response> {
 	let client = reqwest::Client::new();
 
-	let server_secret = util::env::read_secret(&["hcaptcha", "secret"]).await?;
+	let secret_key = if let Some(secret_key) = &ctx.secret_key {
+		secret_key.clone()
+	} else {
+		util::env::read_secret(&["hcaptcha", "secret"]).await?
+	};
 
 	let res = client
 		.post("https://hcaptcha.com/siteverify")
@@ -25,7 +29,7 @@ async fn handle(
 		.body(format!(
 			"response={client_response}&secret={secret}&sitekey={site_key}&remoteip={remote_address}",
 			client_response = ctx.client_response,
-			secret = server_secret,
+			secret = secret_key,
 			site_key = ctx.site_key,
 			remote_address = ctx.remote_address,
 		))